sci-biology/deeptools: Replace git-2 eclass
It looks like the git-2 eclass does not exist (anymore)[1].
I assume it was superseded by the git-r3 eclass[2].
A quick search brought up a commit[3] indicating it can be used as a
simple drop-in replacement.
Also, I came across a list[4] which seems to call out
`sci-biology/deeptools-9999` as an isolated case of using this eclass.

I also updated the copyright and bumped the EAPI version to make
`repoman full` happy.

[1] https://devmanual.gentoo.org/eclass-reference/git-2.eclass/
[2] https://devmanual.gentoo.org/eclass-reference/git-r3.eclass/
[3] https://github.com/wichtounet/wichtounet-overlay/commit/1133fe5b82b2b0728d3ab141ebde9aee88832046
[4] https://gentooqa.levelnine.at/results/science/stats/ECS-STA-ebuild_eclass_statistics/sort-by-filter/git-2.eclass/

Fixes #958.