sci-biology/deeptools: Match ebuilds
The two ebuilds clearly where written to handle both cases, a fixed
version and the lasted git master, based on the ebuild name.
Still, they diverged in EAPI version.

This matches them by symlinnking one to the other.
As the ebuild contains the version number 2.5.0 hardcoded, this
ebuild was chosen as the target of the symlink.